python-engine/pyproject.toml
[project]
name = "flex-analyzer"
version = "0.1.0"
description = "タンパク質の揺らぎ解析エンジン"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "biopython>=1.83",
    "pydantic>=2.5.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
]

[project.scripts]
flex-analyze = "flex_analyzer.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true